% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_object.R
\name{as.data.frame.my_object}
\alias{as.data.frame.my_object}
\title{Convert a my_object to a data frame}
\usage{
\method{as.data.frame}{my_object}(x, ...)
}
\arguments{
\item{x}{An object of class \code{my_object}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data frame with one column named \code{value}.
}
\description{
Convert a my_object to a data frame
}
