% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Produce a map plot of fatalities for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer corresponding to the state number of the state for which a map plot should be drawn}

\item{year}{An integer specifying the year for which the map plot should be drawn}
}
\description{
Produce a map plot of fatalities for a given state and year
}
\note{
An error is thrown if the state number is invalid

A warning is given if given state did not have any fatalities for that year

Requires the map package and dplyr package
}
\examples{
fars_map_state(44,2015)
}
