% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector specifying years of data read}
}
\value{
A list containing a tibble for each year, with only the MONTH and year columns
}
\description{
A helper function to read FARS data for multiple years
}
\note{
Produces a warning if the one of the years passed was invalid, i.e. does not exist in working directory or format of years was incorrect
}
\examples{
fars_read_years(c(2013:2015))

}
