% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Produce summary table of fatalities by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector specifying years of data to summarize}
}
\value{
A summary table giving the number of fatal injuries from crashes, by month and year
}
\description{
A function to summarize the number of fatal injuries from car crashes for a given number of years, broken down by month
}
\examples{
fars_summarize_years(c(2013:2015))

}
