% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create FARS data specific filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character vector specifiying the year to be used in the filename}
}
\value{
A character vector that follows the file name convention for FARS data
}
\description{
A helper function to create a filename that follows the convention for FARS data
}
\examples{
\dontrun{
make_filename(2019)
}


}
