% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_object.R
\name{new_my_object}
\alias{new_my_object}
\title{Constructor for a \code{my_object} S3 class}
\usage{
new_my_object(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An object of class \code{my_object}.
}
\description{
This function creates an object of class \code{my_object} from a numeric vector.
}
