% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_object.R
\name{plot.my_object}
\alias{plot.my_object}
\title{Plot method for my_object}
\usage{
\method{plot}{my_object}(x, ...)
}
\arguments{
\item{x}{An object of class \code{my_object}.}

\item{...}{Additional graphical parameters passed to \code{plot}.}
}
\value{
A plot of the numeric data.
}
\description{
Plot method for my_object
}
