% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_object.R
\name{print.my_object}
\alias{print.my_object}
\title{Print method for my_object}
\usage{
\method{print}{my_object}(x, ...)
}
\arguments{
\item{x}{An object of class \code{my_object}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for my_object
}
