% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_object.R
\name{summary.my_object}
\alias{summary.my_object}
\title{Summary method for my_object}
\usage{
\method{summary}{my_object}(object, ...)
}
\arguments{
\item{object}{An object of class \code{my_object}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A summary of the data inside the object.
}
\description{
Summary method for my_object
}
